+++ tests/testthat/test-extract.R
@@ -39,8 +39,6 @@ test_that("error messages work", {
 test_that("warning messages work", {
         expect_warning(fingertips_data(DomainID = 1938132767, AreaCode = "E06000015", ParentAreaTypeID = 153),
                        "AreaTypeID not a child of ParentAreaTypeID\\. There may be duplicate values in data\\. Use function area_types\\(\\) to see mappings of area type to parent area type\\.")
-        expect_warning(fingertips_data(IndicatorID = 10101, inequalities = F),
-                       "argument inequalities is deprecated; please use categorytype instead\\.")
         expect_warning(fingertips_data(DomainID = 1938133114, ProfileID = 17),
                        "DomainID is complete so ProfileID is ignored")
 })